==> notebooks/02a_differentiate_speakers_sample_approach.ipynb <==
## Speaker classification pipeline
---
This Jupyter notebook contains a pipeline for classifying the speakers in an interview transcript into two categories: "interviewer" and "interviewee". The pipeline uses a reference list of interview questions and calculates matching scores between each row in the transcript and every question in the reference list. The scores are then aggregated at the average level for each speaker, and the speaker with the highest matching score with the questions in the transcript is defined as the "interviewer".

### Pipeline steps

The pipeline consists of the following steps:

1. __Loading the data__: The interview transcript and reference list are loaded into the notebook.

2. __Preprocessing__: The transcript is preprocessed to remove any unnecessary elements, such as timestamps or speaker tags. The reference list is also preprocessed to remove any duplicate questions or irrelevant information.

3. __Calculating matching scores__: For each speaker*i*:  for each speaker *i*'s records in the transcript and each question in the reference list, a matching score is calculated using a text similarity metric. The matching score reflects the degree of similarity between the row and the question. Then for each question we select the maximum score.

4. __Aggregating scores by speaker__: The matching scores are aggregated at the average level for each speaker across questions' scores, creating a summary table that shows the average score for each speaker.

5. __Identifying the interviewer__: The speaker with the highest average score is identified as the "interviewer", while the other speaker(s) are classified as "interviewee(s)".

### Conclusion

This pipeline provides a simple yet effective way of identifying the interviewer in an interview transcript based on the matching scores between the transcript and a reference list of questions. It can be useful for various applications, such as analyzing interview sentiments.

---

import pandas as pd
import difflib  as dl
import numpy as np

### Loading the data

# read a sample transcript which we saved in data frame
transcript = pd.read_excel("../res/processed_transcript_from_json_to_dataframe.xlsx")

transcript.head()

#### Assumptions:
1. In this project we combine spk_1 + spk_2 + spk_3 into one "person" as a "collective interviewee"

transcript['speaker'] = transcript['speaker'].astype(str)

transcript['speaker'] = transcript['speaker'].apply(lambda x:x if x=='spk_0' else 'spk_1')

transcript = transcript[['speaker','text']]
transcript.head(10)

#### This is a reference list of questions, for this interview
The reference list represents high level question for different topics and may varies from all the questioned the interviewer asked. Sometimes interviewer asks more questions for the same topic, or gives some background before asks a question. We will identify and map lines in transcrip to the reference questions and answers to these questions. 

list_of_questions =[
    "From your point of view What is Healthier SG",
    "when you look at our health care system, what was wrong with it? I mean, why are we doing this now? Why are we shifting in this direction?  ",
    "The idea of having family doctors is something so common elsewhere in the world. Why are we only doing it now?",
    "Will this also add on to increasing burdens for GPs? What would Healthier SG mean for GPs?"
    "How dispence of medications will affect your operations?",
    "Fees will be paid to GPS based on the health risk profiles of patients and depending on the outcomes.How are you going to measure this when outcome is going to be so different for everyone?",
    "As a national savings for the healthcare system, will we save money is this is running correctly?",
    "Should the typical patient be concerned about rising costs when they go see the doctor world prices increase over time?",
    "If there's one thing that you could you know change right now fixed right now with regards to our health care system, what would it be?"
]

### Calculating matchin score for *spk_0*


# Select only spk_0 lines:
df_spk0 = transcript[transcript.speaker=="spk_0"]

df_spk0=df_spk0.reset_index()

df_spk0.head()

import difflib  as dl

possibilities = df_spk0.text.tolist()
# get max length for padding:
padding_len = max([len(val) for val in possibilities])
print(padding_len)

def get_similarity(s1,s2, padding_len):
    """
    s1: current line of speaker
    s2: is for question from the reference list of questions
    padding_len: all lines in transcript are aligned to the left side by inserting 
    padding to the right end of the string. This help to ensure a fair comparision.
    """
    # s1 need to be padded with 0:
    s1=s1.ljust(padding_len, '0')
    return dl.SequenceMatcher(None, s2, s1).ratio()
    
    

#### a) Getting a similarity scores for each question vs each line in transcript:

for i,q_i in enumerate(list_of_questions):
    df_spk0[q_i] = 0
    df_spk0[q_i] = df_spk0["text"].apply(lambda x:get_similarity(str(x), q_i,padding_len ))

#### b) For each Reference question (column in our case) calculate the max score and then take the mean:

spk_0_avg_score = np.mean(df_spk0.max().tolist()[3:])
spk_0_avg_score

#### c) create a new column in the transcript and fill out this score as score for the spk_0:

# add this score as a column to the transcript
transcript['spk_0_score']=spk_0_avg_score

### Calculating matchin score for *spk_1*


# Select only interviewer's lines:
df_spk1 = transcript[transcript.speaker=="spk_1"]
df_spk1=df_spk1.reset_index()
df_spk1.head()

#### a) Getting a similarity scores for each question vs each line in transcript:

for i,q_i in enumerate(list_of_questions):
    df_spk1[q_i] = 0
    df_spk1[q_i] = df_spk1["text"].apply(lambda x:get_similarity(str(x), q_i,padding_len ))

df_spk1.head()

#### b) For each Reference question (column in our case) calculate the max score and then take the mean:

spk_1_avg_score = np.mean(df_spk1.max().tolist()[3:])
spk_1_avg_score

#### c) create a new column in the transcript and fill out this score as score for the spk_1:

# add this score as a column to the transcript
transcript['spk_1_score']=spk_1_avg_score

### Identifying the interviewer
Compare the score for spk_0 and spk_1.
Those who has the highest score is "interviewer", because it's the transcript's lines are best matched with the list of reference question

transcript.head(5)

def who_is_interviewer(df,i):
    # if the i-line is for spk_0 and spk_0_score>spk_1_score then spk_0 is interviewer
    if df['speaker'][i]=="spk_0":
        if df["spk_0_score"][i]>=df["spk_1_score"][i]:
            return "interviewer"
        else:
            return "interviewee"
    else: # the line i belongs to spk_1
        if df["spk_0_score"][i]>=df["spk_1_score"][i]:
            return "interviewee"
        else:
            return "interviewer"

# create a new column to keep values who is interviewee/interviewer:
transcript['role']=""
for ind in transcript.index:
    transcript['role'][ind] = who_is_interviewer(transcript, ind)

transcript

### Saving processed results
This results are ready for text processing with AWS 

transcript.to_csv("../res/transcript_with_differentiated_interviewer_interviewee.csv",index=False)

moving to AWS SageMaker
