==> notebooks/02b_mapping_reference_questions_and_answers_in_the_transcript.ipynb <==
---
## Response mapping pipeline

This Jupyter notebook contains a pipeline for mapping lines in an interview transcript to their corresponding reference questions. The pipeline takes into account the fact that the interviewer may ask additional questions related to the reference questions, and that an interviewee's response to a reference question may span multiple lines in the transcript.

### Pipeline Steps

The pipeline consists of the following steps:

1. __Loading the data__: The interview transcript, speaker classification results, and reference question list are loaded into the notebook.

2. __Preprocessing__ (optional): The transcript is preprocessed to remove any unnecessary elements, such as timestamps or speaker tags. The reference question list is also preprocessed to remove any duplicate or irrelevant questions.

3. __Mapping reference questions to interviewer's lines__: For each reference question in the list, the pipeline identifies the line in the transcript that has the highest matching score with the question using a text similarity metric, such as difflib matching score. The identified line is assumed to be an "interviewer's" line and mapped to the corresponding question.

4. __Mapping interviewee's responses to reference questions__: The pipeline then identifies all "interviewee's" lines that occur between two adjacent reference questions and maps them to the first reference question.

5. __Joining responses to the same question__: Finally, the pipeline joins all responses that correspond to the same reference question into a single record, creating a summary table that shows the question and its corresponding responses.

### Conclusion
This pipeline provides a simple yet effective way of mapping lines in an interview transcript to their corresponding reference questions, even in cases where an interviewee's response to a reference question spans multiple lines in the transcript. It can be useful for various applications, such as analyzing the content and structure of interview responses, or identifying patterns in how interviewees answer specific questions.

---

import pandas as pd
import difflib  as dl
import numpy as np

### Loading the data

# read a sample transcript which we saved in data frame
transcript = pd.read_csv("../res/transcript_with_differentiated_interviewer_interviewee.csv")

transcript.head()

transcript['speaker'] = transcript['speaker'].astype(str)

transcript = transcript[['role','text']]
transcript.head(10)

#### This is a reference list of questions, for this interview
The reference list represents high level question for different topics and may varies from all the questioned the interviewer asked. Sometimes interviewer asks more questions for the same topic, or gives some background before asks a question. We will identify and map lines in transcrip to the reference questions and answers to these questions. 

list_of_questions =[
    "From your point of view What is Healthier SG",
    "when you look at our health care system, what was wrong with it? I mean, why are we doing this now? Why are we shifting in this direction?  ",
    "The idea of having family doctors is something so common elsewhere in the world. Why are we only doing it now?",
    "Will this also add on to increasing burdens for GPs? What would Healthier SG mean for GPs?"
    "How dispence of medications will affect your operations?",
    "Fees will be paid to GPS based on the health risk profiles of patients and depending on the outcomes.How are you going to measure this when outcome is going to be so different for everyone?",
    "As a national savings for the healthcare system, will we save money is this is running correctly?",
    "Should the typical patient be concerned about rising costs when they go see the doctor world prices increase over time?",
    "If there's one thing that you could you know change right now fixed right now with regards to our health care system, what would it be?"
]

### Step 1: Mapping reference questions to interviewer's line


# Select only interviewer's lines:
df_spk0 = transcript[transcript.role=="interviewer"]

df_spk0=df_spk0.reset_index()

df_spk0.head()

Some lines in transcript contains a 'reference question, but some are just a part of conversation. So our goal is identify only lines which are the most similar to the reference list. We will use the difflib library for this, and for each interviewer's line and for each question we will calculate the "matching" score. The line with the highest score will be labeled as the line for a reference question.

    
    for each question from the list of questions: 
        create a new column in data frame
        for each line in the interviewer's transcript:
            get the score how similar the line to question
        find the line with the highest matching score
        lable the line with the highest matching score as a line where the reference question was asked
        

import difflib  as dl

possibilities = df_spk0.text.tolist()
# get max length for padding:
padding_len = max([len(val) for val in possibilities])
print(padding_len)

def get_similarity(s1,s2, padding_len):
    """
    s1: current line of speaker
    s2: is for question from the reference list of questions
    padding_len: all lines in transcript are aligned to the left side by inserting 
    padding to the right end of the string. This help to ensure a fair comparision.
    """
    # s1 need to be padded with 0:
    s1=s1.ljust(padding_len, '0')
    return dl.SequenceMatcher(None, s2, s1).ratio()
    
    

Getting a similarity scores for each question vs each line in transcript:

for i,q_i in enumerate(list_of_questions):
    df_spk0[q_i] = 0
    df_spk0[q_i] = df_spk0["text"].apply(lambda x:get_similarity(str(x), q_i,padding_len ))

df_spk0

Map reference question based on max similarity score

# Create a new column
df_spk0['reference_q_id']=""

We will print the reference question and the line in transcript it has been map too

for i,q_i in enumerate(list_of_questions):
    # get row index with max score in each column:
    max_score_index = df_spk0[q_i].idxmax()
    print(q_i)
    print(max_score_index)
    print(df_spk0.at[max_score_index,'text'])
    df_spk0.at[max_score_index,'reference_q_id'] = q_i
    print('-----')

df_spk0=df_spk0[['role','text','reference_q_id']]

### Join our results back to the original transcript

transcript = pd.merge(transcript,df_spk0,how='left',on=['role','text'])

transcript.tail(20)

## Step 2: Mapping interviewee's responses to reference questions


import numpy as np

# 1. Find Index for each questions:
question_index_dict ={}
question_index_list = []
for q_i in list_of_questions:
    
    index_q_i = np.where(transcript['reference_q_id'].isin([q_i]))
    question_index_dict[q_i]=index_q_i[0][0]
    question_index_list.append(index_q_i[0][0])
    

question_index_list

question_index_dict

# Step 2: map answeres 
index_answers = dict()
for i in range(len(question_index_list)-1):
    start = question_index_list[i]+1
    end = question_index_list[i+1]
    index_answers[list_of_questions[i]]=list(range(start,end))
# last question and answers:
index_answers[list_of_questions[-1]] = list(range(question_index_list[-1]+1,transcript.shape[0]))


transcript.shape[0]


transcript['responce_to_question']='empty'

for q_i in list_of_questions:
    response_indexes = index_answers.get(q_i)
    for i in response_indexes:
        transcript.at[i,'responce_to_question'] = q_i

transcript.head(10)

## Step 3: Joining responses to the same question

# Select only speaker 1 , and all text replies , except with 'empty' values in the responce_to_question
transcript_responses = transcript[transcript.role=='interviewee'][['text','responce_to_question']]

transcript_responses=transcript_responses[transcript_responses.responce_to_question!='empty']

transcript_responses_agg = transcript_responses.groupby(['responce_to_question'])['text'].apply(lambda x: ','.join(x)).reset_index()

# drop all ',':
transcript_responses_agg['responce_to_question'] = transcript_responses_agg['responce_to_question'].apply(
    lambda x:x.replace(',',''))
transcript_responses_agg['text'] = transcript_responses_agg['text'].apply(
    lambda x:x.replace(',',''))

transcript_responses_agg.head(20)

transcript['responce_to_question'] = transcript_responses_agg['responce_to_question'].apply(
    lambda x:x.replace(',',''))
transcript['text'] = transcript_responses_agg['text'].apply(
    lambda x:x.replace(',',''))

### Saving processed results
This results are ready for text processing with AWS 

transcript.to_csv("../res/transcript_with_mapped_questions_and_answers_all.csv",index=False)
transcript_responses_agg.to_csv("../res/transcript_with_mapped_questions_and_answers.csv", index=False)

moving to AWS SageMaker
