==> notebooks/03_topic_modelling_with_aws_comprehend.ipynb <==
---
## Topic Modeling Pipeline with AWS Comprehend

This Jupyter notebook implements a pipeline for topic modeling using AWS Comprehend. The pipeline takes as input a file containing interview questions and responses, and generates topics and corresponding keywords using the Comprehend topic modeling API.

The pipeline consists of the following steps:

1. __Data Preprocessing__: The input file is read and preprocessed to remove any unwanted characters or formatting. This is done using Python's pandas and re libraries.

2. __Calling AWS Comprehend Topic Modeling API__: The preprocessed data is then passed to the Comprehend topic modeling API using the AWS SDK for Python (boto3). The API generates topics and corresponding keywords based on the input data.

3. __Extracting Results__: The generated results are then extracted and stored in a pandas dataframe for further sentiment analysis.

This pipeline can be customized to fit different types of input data and analysis requirements.

---

### Set up environment:

#!pip install sagemaker ipywidgets --upgrade --quiet

import time
import boto3
from botocore.exceptions import ClientError
import requests

comprehend = boto3.client('comprehend', region_name='us-east-1')

import pandas as pd
import os

### Read input file with interview questions and responses

BUCKET='cnatest' # Or whatever you called your bucket
data_key = 'transcript_with_mapped_questions_and_answers.csv' # Where the file is within your bucket
data_location = 's3://{}/{}'.format(BUCKET, data_key)
df = pd.read_csv(data_location)

# Transformed review, input for Comprehend
LOCAL_TRANSFORMED_INTERVIEW = os.path.join('data', 'Transformed.txt')
S3_OUT = 's3://' + BUCKET + '/out/' + 'Transformed.txt'

# Final dataframe where topics and sentiments are going to be joined
S3_TOPICS = 's3://' + BUCKET + '/out/' + 'FinalDataframe.csv'

df.head()

### Cleaning responses

def clean_text(df):
    """Preprocessing review text.
    The text becomes Comprehend compatible as a result.
    This is the most important preprocessing step.
    """
    # Encode and decode reviews
    df['text'] = df['text'].str.encode("utf-8", "ignore")
    df['text'] = df['text'].str.decode('ascii')
    
    df['text'] = df['text'].str.replace('[PII]','', regex=True)
    # Replacing characters with whitespace
    df['text'] = df['text'].replace(r'\r+|\n+|\t+|\u2028',' ', regex=True)

    # Replacing punctuations
    df['text'] = df['text'].str.replace('[^\w\s]','', regex=True)

    # Lowercasing reviews
    df['text'] = df['text'].str.lower()
    return df
df = clean_text(df)
df.head(10)

def prepare_input_data(df):
    """Encoding and getting reviews in byte size.
    Review gets encoded to utf-8 format and getting the size of the reviews in bytes. 
    Comprehend requires each review input to be no more than 5000 Bytes
    """
    df['textsize'] = df['text'].apply(lambda x:len(x.encode('utf-8')))
    df = df[(df['textsize'] > 0) & (df['textsize'] < 5000)]
    df = df.drop(columns=['textsize'])
    return df
df = prepare_input_data(df)

# We first save the input file locally
with open(LOCAL_TRANSFORMED_INTERVIEW, "w") as outfile:
    outfile.write("\n".join(df['text'].tolist()))
    

### Run an Amazon Comprehend topic modeling job

# Client and session information
session = boto3.Session()
s3 = boto3.resource('s3')

# Account id. Required downstream.
account_id = boto3.client('sts').get_caller_identity().get('Account')

# Initializing Comprehend client
comprehend = boto3.client(service_name='comprehend', 
                          region_name=session.region_name)


# Number of topics set to 5 after having a human-in-the-loop
# This needs to be fully aligned with topicMaps dictionary in the third script 
NUMBER_OF_TOPICS = 5

# Input file format of one review per line
input_doc_format = "ONE_DOC_PER_LINE"

# Role arn (Hard coded, masked)
data_access_role_arn = ""



# Constants for S3 bucket and input data file
BUCKET='cnatest' 
input_s3_url = 's3://' + BUCKET + '/out/' + 'Transformed.txt'
output_s3_url = 's3://' + BUCKET + '/out/' + 'output/'

# Final dataframe where we will join Comprehend outputs later
S3_FEEDBACK_TOPICS = 's3://' + BUCKET + '/out/' + 'FinalDataframe.csv'

# Local copy of Comprehend output
LOCAL_COMPREHEND_OUTPUT_DIR = os.path.join('comprehend_out', '')
LOCAL_COMPREHEND_OUTPUT_FILE = os.path.join(LOCAL_COMPREHEND_OUTPUT_DIR, 'output.tar.gz')

INPUT_CONFIG={
    # The S3 URI where Comprehend input is placed.
    'S3Uri':    input_s3_url,
    # Document format
    'InputFormat': input_doc_format,
}
OUTPUT_CONFIG={
    # The S3 URI where Comprehend output is placed.
    'S3Uri':    output_s3_url,
}

import json
# Start Comprehend topic modelling job.
# Specifies the number of topics, input and output config and IAM role ARN 
# that grants Amazon Comprehend read access to data.
start_topics_detection_job_result = comprehend.start_topics_detection_job(
                                                    NumberOfTopics=NUMBER_OF_TOPICS,
                                                    InputDataConfig=INPUT_CONFIG,
                                                    OutputDataConfig=OUTPUT_CONFIG,
DataAccessRoleArn=data_access_role_arn)

print('start_topics_detection_job_result' )

# Job ID is required downstream for extracting the Comprehend results
job_id = start_topics_detection_job_result["JobId"]
#print('job_id: ', job_id)


# Keeping track if Comprehend has finished its job
description = comprehend.describe_topics_detection_job(JobId=job_id)

topic_detection_job_status = description['TopicsDetectionJobProperties']["JobStatus"]
print(topic_detection_job_status)
while topic_detection_job_status not in ["COMPLETED", "FAILED"]:
    time.sleep(120)
    topic_detection_job_status = comprehend.describe_topics_detection_job(JobId=job_id)['TopicsDetectionJobProperties']["JobStatus"]
    print(topic_detection_job_status)

topic_detection_job_status = comprehend.describe_topics_detection_job(JobId=job_id)['TopicsDetectionJobProperties']["JobStatus"]
print(topic_detection_job_status)

comprehend.describe_topics_detection_job(JobId=job_id)

### Extracting Results:

When the job is successfully complete, it returns a compressed archive containing two files: topic-terms.csv and doc-topics.csv. The first output file, topic-terms.csv, is a list of topics in the collection. For each topic, the list includes, by default, the top terms by topic according to their weight. The second file, doc-topics.csv, lists the documents associated with a topic and the proportion of the document that is concerned with the topic. 

The outputs of Amazon Comprehend are copied locally for our next steps.

# Bucket prefix where model artifacts are stored
prefix = f'{account_id}-TOPICS-{job_id}'

# Model artifact zipped file
artifact_file = 'output.tar.gz'

# Location on S3 where model artifacts are stored
target = f's3://{BUCKET}/out/output/{prefix}/output/{artifact_file}'





# Copy Comprehend output from S3 to local notebook instance
! aws s3 cp {target}  ./comprehend-out/

### Extract output file and read them with pandas

import tarfile

# Unzip the Comprehend output file. 
# Two files are now saved locally- 
#       (1) comprehend-out/doc-topics.csv and 
#       (2) comprehend-out/topic-terms.csv

comprehend_tars = tarfile.open('comprehend-out/output.tar.gz')
comprehend_tars.extractall('comprehend-out')
comprehend_tars.close()

df_topics = pd.read_csv('comprehend-out/topic-terms.csv')

df_docs_by_topics = pd.read_csv('comprehend-out/doc-topics.csv')

df_topics.head(3)

df_docs_by_topics.head(20)

### Overall we have 5 topics. Lets look closer to terms:

for i in range(5):
    df_tmp = df_topics[df_topics.topic==i]
    print(df_tmp.head(20))

### Topic names for 5 topics created by human-in-the-loop or SME feed

topicMaps = {
    0: 'Cost',
    1: 'Managment of Chronic Conditions',
    2: 'System Operation',
    3: 'Patient Experience',
    4: 'Changes in Healthcare',
}

Proceed to the notebook _"04_topic_mapping_sentiment.ipynb"_.
