==> notebooks/04_topic_mapping_sentiment.ipynb <==
---
## A Pipeline to Understand Topics and Sentiment

We may encounter multiple individuals from diverse demographics undergoing an interview in a real-life scenario. Therefore, on top of generated topics, we can build a pipeline for sentiment analysis to find how interviewees  are reacting to questions, and also analyze aggregated information on user affinity towards a particular question.

The pipeline consists of the following steps:
1. __Data Preprocessing__: The input file is read and preprocessed to remove any unwanted characters or formatting. This is done using Python's pandas and re libraries.
2. __Name the topics__: use a human-in-the-loop with enough domain knowledge and subject matter expertise to name the topics by looking at their associated terms.
3. __Calling AWS Comprehend Detect Sentiment API__: The preprocessed data is then passed to the Comprehend detect sentiment API using the AWS SDK for Python (boto3). The API generates sentiment for each "mock-inteviewee's" each question answer text. 
4. __Aggregating topics and sentiments__: Both topics and sentiments are tightly coupled with answers. We will be aggregating topics and sentiments at question level and count the composite keys for each question. This final step helps us better understand the granularity of the answers per questions and categorizing it per topic in an aggregated manner.

This analysis is inspired by this [example](https://aws.amazon.com/blogs/machine-learning/get-better-insight-from-reviews-using-amazon-comprehend/)

---


import time
import boto3
from botocore.exceptions import ClientError
import requests
import pandas as pd

session = boto3.Session()
comprehend = boto3.client(  'comprehend')

### Step 1-2: Data Preprocessing and Name the topics

topicMaps = {
    0: 'Cost',
    1: 'Managment of Chronic Conditions',
    2: 'System Operation',
    3: 'Patient Experience',
    4: 'Changes in Healthcare',
}

topic_map_df = pd.DataFrame(topicMaps.items())
topic_map_df.columns=['topic','topic_name']
topic_map_df.head(3)

BUCKET='cnatest'
# Final dataframe where we will join Comprehend outputs later
S3_FEEDBACK_TOPICS = 's3://' + BUCKET + '/out/' + 'FinalDataframe.csv'
# Final output
S3_FINAL_OUTPUT = 's3://' + BUCKET + '/out/' + 'TopicsSentiments.csv'
language_code = 'en'
TOP_TOPICS=3

# Loading documents and topics assigned to each of them by Comprehend
DOC_TOPIC_FILE='comprehend-out/doc-topics.csv'
docTopics = pd.read_csv(DOC_TOPIC_FILE)
docTopics.head()

# Creating a field with doc number. 
# This doc number is the line number of the input file to Comprehend.
docTopics['doc'] = docTopics['docname'].str.split(':').str[1]
docTopics['doc'] = docTopics['doc'].astype(int)
docTopics.head()

# Load topics and associated terms
DOC_TOPIC_FILE='comprehend-out/topic-terms.csv'
topicTerms = pd.read_csv(DOC_TOPIC_FILE)
# Consolidate terms for each topic
aggregatedTerms = topicTerms.groupby('topic')['term'].aggregate(lambda term: term.unique().tolist()).reset_index()
# Sneak peek
aggregatedTerms = pd.merge(aggregatedTerms, topic_map_df, on='topic',how='left')
aggregatedTerms.head()

### We will use the same file with responses to simulate that we have multiple interviewees
1. We read the same input file 3 times
2. Concatenate them into one dataframe
3. Join with the results from the topic modelling
4. Then for each reply we will generate sentiment 
5. Finally we aggregate the data to question level 

data_key = 'transcript_with_mapped_questions_and_answers.csv' # Where the file is within your bucket
data_location = 's3://{}/{}'.format(BUCKET, data_key)

interviewee_1 = pd.read_csv(data_location)
interviewee_1 = interviewee_1.reset_index()
interviewee_2 = pd.read_csv(data_location)
interviewee_2 = interviewee_2.reset_index()
interviewee_3 = pd.read_csv(data_location)
interviewee_3 = interviewee_3.reset_index()
interviewee_1.head(10)

df_mock = interviewee_1.append(interviewee_2)
df_mock = df_mock.append(interviewee_3)
df_mock.shape

df_mock.columns=['doc', 'question', 'text']

def clean_text(df):
    """Preprocessing review text.
    The text becomes Comprehend compatible as a result.
    This is the most important preprocessing step.
    """
    # Encode and decode reviews
    df['text'] = df['text'].str.encode("utf-8", "ignore")
    df['text'] = df['text'].str.decode('ascii')
    
    df['text'] = df['text'].str.replace('[PII]','', regex=True)
    # Replacing characters with whitespace
    df['text'] = df['text'].replace(r'\r+|\n+|\t+|\u2028',' ', regex=True)

    # Replacing punctuations
    df['text'] = df['text'].str.replace('[^\w\s]','', regex=True)

    # Lowercasing reviews
    df['text'] = df['text'].str.lower()
    return df
df_mock = clean_text(df_mock)
def prepare_input_data(df):
    """Encoding and getting reviews in byte size.
    Review gets encoded to utf-8 format and getting the size of the reviews in bytes. 
    Comprehend requires each review input to be no more than 5000 Bytes
    """
    df['textsize'] = df['text'].apply(lambda x:len(x.encode('utf-8')))
    
    df = df[(df['textsize'] > 0) & (df['textsize'] < 5000)]
  
    df = df.drop(columns=['textsize'])
    return df
df_mock = prepare_input_data(df_mock)


# The index of feedbackTopics is referring to doc field of docTopics dataframe
repliesTopics = pd.merge(df_mock, 
                          docTopics, 
                          on='doc', 
                          how='left')

repliesTopics.head(15)

# Reviews will now have topic numbers, associated terms and topics names
repliesTopics = repliesTopics.merge(aggregatedTerms, 
                                      on='topic', 
                                      how='left')
repliesTopics.head()

### Step 3: Generate sentiment for the replies text using detect_sentiment. 
It inspects text and returns an inference of the prevailing sentiment (POSITIVE, NEUTRAL, MIXED, or NEGATIVE).

def detect_sentiment(text, language_code):
    """Detects sentiment for a given text and language
    """
    comprehend_json_out = comprehend.detect_sentiment(Text=text, LanguageCode=language_code)
    return comprehend_json_out

# Comprehend output for sentiment in raw json 
repliesTopics['comprehend_sentiment_json_out'] = repliesTopics['text'].apply(lambda x: detect_sentiment(x, language_code))

# Extracting the exact sentiment from raw Comprehend Json
repliesTopics['sentiment'] = repliesTopics['comprehend_sentiment_json_out'].apply(lambda x: x['Sentiment'])

# Sneak peek
repliesTopics.head(2)

# Creating a composite key of topic name and sentiment.
# This is because we are counting frequency of this combination.
repliesTopics['TopicSentiment'] = repliesTopics['topic_name'] + '_' + repliesTopics['sentiment']

repliesTopics.head(3)

repliesTopics['sentiment'].value_counts()

repliesTopics['TopicSentiment'].value_counts()

### Step 4: Aggregating topics and sentiments
This final step helps us better understand the granularity of the replies per question 

# Create question id group
questions_DF = repliesTopics.groupby('question')

questions_DF

# Each product now has a list of topics and sentiment combo (topics can appear multiple times)
topicDF = questions_DF['TopicSentiment'].apply(lambda x:list(x)).reset_index()
topicDF.head()

# Count appreances of topics-sentiment combo for product
from collections import Counter

topicDF['TopTopics'] = topicDF['TopicSentiment'].apply(Counter)
topicDF.head()

# Sorting topics-sentiment combo based on their appearance
topicDF['TopTopics'] = topicDF['TopTopics'].apply(lambda x: sorted(x, key=x.get, reverse=True))

# Select Top k topics-sentiment combo for each product/review
topicDF['TopTopics'] = topicDF['TopTopics'].apply(lambda x: x[:TOP_TOPICS])

# Sneak peek
topicDF.head()

for q in topicDF.question.tolist():
    print(f'QUESTION: {q}')
    print('TOPICS:')
    [print(val) for val in topicDF[topicDF.question==q].TopTopics.tolist()[0]]
    print('---')

### Saving results:

topicDF.to_csv(S3_FINAL_OUTPUT,index=False)
