==> notebooks/05_sentiment_analysis_entety_recognition_with_aws.ipynb <==
---
## Key phrase sentiment analysis pipeline

This Jupyter notebook contains a pipeline for extracting targeted sentiment for key phrases in interview responses using AWS Comprehend. The pipeline first identifies the key phrases in the responses using the AWS Comprehend Key Phrases Detect API, and then extracts the context surrounding each key phrase to provide additional context for sentiment analysis. The pipeline can be used to identify patterns in how interviewees express positive or negative sentiment towards specific key phrase.

### Pipeline steps
The pipeline consists of the following steps:

1. __Loading the data__: The interview transcript, speaker classification results, and reference question list are loaded into the notebook.

2. __Preprocessing__: The transcript is preprocessed to remove any unnecessary elements, such as timestamps or speaker tags.

3. __Identifying key phrases__: The pipeline calls the AWS Comprehend Key Phrases Detect API for each response in the transcript, and selects only the top key phrases with a score above 0.99. These key phrases are then used as the basis for sentiment analysis.

4. __Extracting context__: For each top key phrase, the pipeline extracts the context surrounding the phrase by taking 10 words before and after the phrase in the response. This provides additional context for sentiment analysis and helps to ensure that the sentiment is targeted specifically to the key phrase.

5. __Detecting sentiment__: The pipeline calls the AWS Comprehend Detect Sentiment API for each extracted context, and records the sentiment score and sentiment label (positive, negative, neutral, or mixed) for each key phrase.

6. __Aggregating results__: The pipeline can be used to aggregate the sentiment analysis results across multiple interviewees and/or questions, providing a more comprehensive understanding of the sentiment towards specific topics or themes.

### Conclusion
This pipeline provides a powerful tool for analyzing the sentiment expressed in interview responses. By identifying key phrases and extracting the context surrounding them, the pipeline enables targeted sentiment analysis that can reveal patterns and insights that might otherwise be missed.

---


import time
import boto3
from botocore.exceptions import ClientError
import requests
import pandas as pd
import os

comprehend = boto3.client('comprehend', region_name='us-east-1')

### Step 1: Loading the data

BUCKET='cnatest' # Or whatever you called your bucket
data_key = 'transcript_with_mapped_questions_and_answers.csv' # Where the file is within your bucket
data_location = 's3://{}/{}'.format(BUCKET, data_key)
df = pd.read_csv(data_location)

df.head(10)

list_questions = df.responce_to_question.tolist()

### Step 2: Preprocessing

def clean_text(df):
    """Preprocessing review text.
    The text becomes Comprehend compatible as a result.
    This is the most important preprocessing step.
    """
    # Encode and decode reviews
    df['text'] = df['text'].str.encode("utf-8", "ignore")
    df['text'] = df['text'].str.decode('ascii')
    
    df['text'] = df['text'].str.replace('[PII]','', regex=True)
    # Replacing characters with whitespace
    df['text'] = df['text'].replace(r'\r+|\n+|\t+|\u2028',' ', regex=True)

    # Replacing punctuations
    df['text'] = df['text'].str.replace('[^\w\s]','', regex=True)

    # Lowercasing reviews
    df['text'] = df['text'].str.lower()
    return df
df = clean_text(df)


### Step 3: Identifying key phrases

def get_insights(sample):
    # Key phrases
    phrases = comprehend.detect_key_phrases(Text=sample, LanguageCode='en')  
    
    return {'phrases':phrases}

keyphrases_dict = dict()

for answer_i, question_i in zip(df.text.tolist(), list_questions):
    keyphrases_dict[question_i] = get_insights(answer_i)
    

keyphrases_dict[df.responce_to_question.tolist()[0]].get('phrases').get('KeyPhrases')[0]

### Save Results of all KeyPhrases to data frame:

key_phrases_df = pd.DataFrame()
for q_i in df.responce_to_question.tolist():
    tmp = pd.DataFrame(keyphrases_dict[q_i].get('phrases').get('KeyPhrases'))
    tmp['question'] = q_i
    key_phrases_df = key_phrases_df.append(tmp)

key_phrases_df.head()

####  Consider only TOP key phrases whith confidence score >0.98

top_key_phrases = key_phrases_df[key_phrases_df.Score>0.98]

set_of_top_key_phrases = set(top_key_phrases.Text.tolist())

len(set_of_top_key_phrases)

### Detect overall sentiment for answers:


def get_insights_sentiment(sample):
    # Key phrases
    sentiment = comprehend.detect_sentiment(Text=sample, LanguageCode='en')  
    
    return {'sentiment':sentiment}

def get_top_500_charackters(input_str, textsize):
    if textsize<5000:
        return input_str
    else:
        return input_str[:500]
    

def prepare_input_data(df):
    """Encoding and getting reviews in byte size.
    Review gets encoded to utf-8 format and getting the size of the reviews in bytes. 
    Comprehend requires each review input to be no more than 5000 Bytes
    """
    df['textsize'] = df['text'].apply(lambda x:len(x.encode('utf-8')))
    df['text_for_sentiment'] = ""
    df['text_for_sentiment'] = df[['text','textsize']].apply(lambda x:get_top_500_charackters(x[0],x[1]),axis=1)
    #df = df[(df['textsize'] > 0) & (df['textsize'] < 5000)]
    df = df.drop(columns=['textsize'])
    return df
df = prepare_input_data(df)

df.head()

sentiment_dict = dict()

for answer_i, question_i in zip(df.text_for_sentiment.tolist(), list_questions):
    sentiment_dict[question_i] = get_insights_sentiment(answer_i)

q_i='As a national savings for the healthcare system will we save money is this is running correctly?'
sentiment_dict[q_i].get('sentiment').get('Sentiment')

sentiment_dict[q_i].get('sentiment')

### Add a column with overall sentiment to input transcript

df.head()

import numpy as np

df['overall_sentiment']=""
for q_i in list_questions:
    sentiment_i = sentiment_dict[q_i].get('sentiment').get('Sentiment')
    index_q_i = np.where(df['responce_to_question']==q_i)
    df.at[index_q_i[0][0],'overall_sentiment'] = sentiment_i

df.head(10)

[ len(val) for val in df.text.tolist()]

### Step 4: Extracting context

### Analyse sentiment for top key phrases
We will consider a context +- 10 words around the "key phrase"

top_key_phrases.columns = ['Score', 'Text', 'BeginOffset', 'EndOffset', 'responce_to_question']

df_sentiment_top_key_phrases = pd.merge(top_key_phrases,df,on="responce_to_question", how="left")

df_sentiment_top_key_phrases.head()

def get_top_substring_10_words(BeginOffset,input_str):
    top = input_str[:BeginOffset]
    words = top.split(" ")
    if len(words)>=10:
        # return last 10
        return ' '.join(words[-10:])
    else:
        return ' '.join(words) 
    
def get_last_substring_10_words(EndOffset,input_str):
    top = input_str[EndOffset+1:]
    words = top.split(" ")
    if len(words)>=10:
        # return last 10
        return ' '.join(words[:10])
    else:
        return  ' '.join(words)  

df_sentiment_top_key_phrases["top_10_words"]=""
df_sentiment_top_key_phrases["bottom_10_words"]=""
df_sentiment_top_key_phrases["context"]=""

df_sentiment_top_key_phrases["top_10_words"]=df_sentiment_top_key_phrases[["BeginOffset","text"]].apply(
lambda x: get_top_substring_10_words(x[0],x[1]),axis=1)
df_sentiment_top_key_phrases["bottom_10_words"]=df_sentiment_top_key_phrases[["EndOffset","text"]].apply(
lambda x: get_last_substring_10_words(x[0],x[1]),axis=1)

def concat_str(top_5_words,Text,bottom_5_words):
    return str(top_5_words)+" "+str(Text)+" "+str(bottom_5_words)

df_sentiment_top_key_phrases["context"]=df_sentiment_top_key_phrases[["top_10_words",
                                                                      "Text",
                                                                     "bottom_10_words"
                                                                     ]].apply(lambda x:concat_str(x[0],x[1],x[0]),axis=1)

df_sentiment_top_key_phrases.head()

df_sentiment_top_key_phrases.shape

### Step 5: Detecting sentiment for key phrases:

context = df_sentiment_top_key_phrases.context.tolist()

sentiment_dict = dict()
key_phrases_list = df_sentiment_top_key_phrases.Text.tolist()
ind = 0
for phrase_i, context_i in zip(key_phrases_list, context):
    sentiment_dict[ind] = get_insights_sentiment(context_i)
    ind = ind+1

df_sentiment_top_key_phrases =df_sentiment_top_key_phrases[['Score', 'Text', 'BeginOffset', 'EndOffset', 'responce_to_question',
       'text', 'overall_sentiment','context']]

for i, phrase in enumerate(key_phrases_list):
    sentiment_i = sentiment_dict[i].get('sentiment').get('Sentiment')
    df_sentiment_top_key_phrases.at[i,'key_phrase_sentiment'] = sentiment_i

df_sentiment_top_key_phrases[df_sentiment_top_key_phrases.key_phrase_sentiment == "NEGATIVE"].head(30)

### Saving Results:

df_sentiment_top_key_phrases.to_csv("data/key_phrases_sentiment.csv", index=False)

df.to_csv("data/overall_sentiment_for_question_replies.csv", index=False)

key_phrases_df.to_csv("data/all_key_phrases.csv",index=False)
